requirements.txt
edge_tts==6.1.14
Flask==3.0.3
grpcio==1.67.0
keyboard==0.13.5
protobuf==5.28.3
PyAudio==0.2.14
pydub==0.25.1
SpeechRecognition==3.10.4